requirements.txt
Django==4.2.0
djangorestframework==3.14.0
joblib==1.5.3
lxml==6.0.2
numpy==2.4.2
pandas==3.0.0
PyPDF2==3.0.1
python-docx==1.2.0
scikit-learn>=1.4.0
lightgbm>=4.0.0
requests>=2.31.0
catboost>=1.2
transformers>=4.30.0
torch>=2.0.0
groq>=0.5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
